WIP: Add commit game log post
The commit game log post prepares the log file to be used by
later processes (dispatcher, logger and machine-manager/emulator).
It saves the log in a json file named after the tournament id, compresses
it, archives it and truncates it to the conventioned log file drive of
1MB. Once the file is prepared, the merkle tree root hash of the file
is calculated and sent in the post payload to the dispatcher.

Also added initial application logging support