Update tournament mapper
Tournament mapper is compliant with the current anuto dapp
dispatcher format, but there is still some info that is
not available, like the tournament champion.

Updated dispatcher api and contract helper classes and
created a tournament info mapping file to get
tournament names while that info is not available
through the blockchain.

Also created an util class to recover all tournaments
properly mapped.